// packages/frontend/src/generated/types/DispatchItemResponse.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

/**
 * @description Response DTO for the `dispatch_item` entity.
 */
export interface DispatchItemResponse {
  /**
   * @type string
   */
  createdAt: string
  /**
   * @type string, uuid
   */
  createdBy: string
  /**
   * @type string,null
   */
  deletedAt?: string | null
  /**
   * @type string,null, uuid
   */
  deletedBy?: string | null
  /**
   * @type string, uuid
   */
  dispatchDocId: string
  /**
   * @type string
   */
  dispatchedAmount: string
  /**
   * @type string, uuid
   */
  id: string
  /**
   * @type string, uuid
   */
  originDbId: string
  /**
   * @type string, uuid
   */
  productId: string
  /**
   * @type string, uuid
   */
  storageId: string
  /**
   * @type string
   */
  updatedAt: string
  /**
   * @type string, uuid
   */
  updatedBy: string
}